Moves remapScrollingPlane and isMouseWheelYEvent into Overflow.utils
