+++ src/components/Overflow/Overflow.jsx
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom'
 import getValidProps from '@helpscout/react-utils/dist/getValidProps'
 import EventListener from '../EventListener'
 import classNames from 'classnames'
-import { remapScrollingPlane } from '../../utilities/scrolling'
+import { remapScrollingPlane } from './Overflow.utils'
 import { smoothScrollTo } from '../../utilities/smoothScroll'
 import { hasContentOverflowX } from '../../utilities/node'
 import {
